Initialize Kubebuilder Project for 'HelloWorld' Operato
This commit initializes the 'HelloWorld' operator project using Kubebuilder's scaffolding tool.
The Kubebuilder command used for project setup is:
kubebuilder init --domain example.com --repo helloworld-operator

The --domain flag specifies the domain associated with the API group for custom resources, which in this case is set as 'example.com'.
The --repo flag defines the name of the repository to be created, named 'helloworld-operator'.

This setup lays the foundation for building a Kubernetes operator that manages 'HelloWorld' custom resources within the Kubernetes cluster.